requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
fastapi==0.129.0
flatbuffers==25.12.19
gast==0.7.0
google-pasta==0.2.0
grpcio==1.78.1
h11==0.16.0
h5py==3.15.1
idna==3.11
keras==3.13.2
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.4.2
opt_einsum==3.4.0
optree==0.18.0
packaging==26.0
pillow==12.1.1
protobuf==6.33.5
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyMySQL==1.1.2
python-dotenv==1.2.1
python-multipart==0.0.22
requests==2.32.5
rich==14.3.3
setuptools==82.0.0
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.52.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
Werkzeug==3.1.6
wheel==0.46.3
wrapt==2.1.1
